feat : filtering done
