+++ README.md
@@ -16,13 +16,13 @@
 ## 🎯🎯 심화 요구사항
 
 - [ ] 1. 데이터를 불러오기전 로딩바를 이용해, 사용자가 데이터가 불러와지고 있다는 것을 보여줍니다.
-- [ ] 2. fetch api 사용하는 부분을 async await을 사용하여 리팩토링합니다.
-- [ ] 3. github issue에서 라벨을 붙이는 것처럼, 우선순위에 따라서 badge를 추가합니다.
-- [ ] 4. ES6 impot & export를 이용해 자바스크립트 파일을 리팩토링합니다.
+- [x] 2. fetch api 사용하는 부분을 async await을 사용하여 리팩토링합니다.
+- [x] 3. github issue에서 라벨을 붙이는 것처럼, 우선순위에 따라서 badge를 추가합니다.
+- [x] 4. ES6 impot & export를 이용해 자바스크립트 파일을 리팩토링합니다.
 
 ## 🕵️‍♂️ 제약사항
 
-- [ ] 1. User의 이름은 최소 2글자 이상이어야 합니다.
+- [x] 1. User의 이름은 최소 2글자 이상이어야 합니다.
 
 <br/><br/>
